+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.1.0 - 2024-06-09
+
+This has now been tested with jetstream2 and should be able to create a cluster on jetstream2.
+
+## Added
+
+- Added configuration for manila-nfs and manila-cephs
+- Can set region from default
+- Can set os images
+
 ## 3.0.0 - 2024-02-21
 
 Removed all deprecated code. Clusters should now be defined using cluster.json.